requirements.txt
beautifulsoup4==4.6.0
certifi==2017.7.27.1
chardet==3.0.4
future==0.16.0
httplib2==0.10.3
idna==2.6
oauth2==1.9.0.post1
oauthlib==2.0.2
python-twitter==3.3
requests==2.18.4
requests-oauthlib==0.8.0
urllib3==1.22
wikipedia==1.4.0
